+++ build_tools/circle/build_test_pypy.sh
@@ -18,7 +18,8 @@ which python
 
 pip install --extra-index https://antocuni.github.io/pypy-wheels/ubuntu numpy Cython pytest
 pip install "scipy>=1.1.0"
-pip install scikit-learn pandas statsmodels
+pip install "scikit-learn==0.19.1"
+pip install pandas statsmodels matplotlib
 
 ccache -M 512M
 export CCACHE_COMPRESS=1