+++ build_tools/travis/install.sh
@@ -49,14 +49,14 @@ if [[ "$DISTRIB" == "conda" ]]; then
             numpy=$NUMPY_VERSION scipy=$SCIPY_VERSION \
             mkl cython=$CYTHON_VERSION \
             scikit-learn=$SCIKIT_LEARN_VERSION \
-            statsmodels=$STATSMODELS_VERSION pytest pytest-cov
+            statsmodels=$STATSMODELS_VERSION
 
     else
         conda create -n testenv --yes python=$PYTHON_VERSION pip nose pytest \
             numpy=$NUMPY_VERSION scipy=$SCIPY_VERSION \
             nomkl cython=$CYTHON_VERSION \
             scikit-learn=$SCIKIT_LEARN_VERSION \
-            statsmodels=$STATSMODELS_VERSION pytest pytest-cov
+            statsmodels=$STATSMODELS_VERSION
     fi
     source activate testenv
 
@@ -76,11 +76,6 @@ if [[ "$DISTRIB" == "conda" ]]; then
 else
     echo "We are only building with DISTRIB=conda currently"
     exit 10
-fi
-
-# use PIP for installing coverage tools since we might not be using a conda dist
-if [[ "$COVERAGE" == "true" ]]; then
-    pip install coverage codecov
 fi
 
 # now run the python setup. This implicitly builds all the C code with build_ext