+++ pmdarima/__init__.py
@@ -6,7 +6,7 @@
 
 import os as _os
 
-__version__ = "1.0.0"
+__version__ = "1.1.0-dev0"
 
 try:
     # this var is injected in the setup build to enable
@@ -55,12 +55,14 @@ else:
 
     # On first import, check the cache, warn if needed
     from ._config import _warn_for_cache_size
+
     _warn_for_cache_size()
 
     # Delete unwanted variables from global
     del _os
     # del _config  # don't delete in case user wants to amend it at top level
     del _warn_for_cache_size
+    # del _config  # don't delete in case user wants to amend it at top level
     del __check_build
     del __PMDARIMA_SETUP__