+++ doc/whats_new.rst
@@ -7,6 +7,25 @@ What's new in pmdarima
 As new releases of pmdarima are pushed out, the following list (introduced in
 v0.8.1) will document the latest features.
 
+v1.1.0
+------
+
+* Add ``ARIMA.plot_diagnostics`` method
+
+* Add new arg to ``ARIMA`` constructor and ``auto_arima``: ``with_intercept`` (default is True).
+
+* New default for ``trend`` is no longer ``'c'``, it is ``None``.
+
+* Added ``to_dict`` method to ``ARIMA`` class to address `Issue 54 <https://github.com/tgsmith61591/pmdarima/issues/54>`_
+
+* ARIMA serialization no longer stores statsmodels results wrappers in the cache,
+  but bundles them into the pickle file. This solves `Issue #48 <https://github.com/tgsmith61591/pmdarima/issues/48>`_
+  and only works on statsmodels 0.9.0+ since they've fixed a bug on their end.
+
+* The ``'PMDARIMA_CACHE'`` and ``'PMDARIMA_CACHE_WARN_SIZE'`` environment variables are
+  now deprecated, since they no longer need to be used.
+
+
 v1.0.0
 ------